working idefics config!
