+++ .env.template
@@ -129,6 +129,7 @@ MODELS=`[
       "websiteUrl": "https://mistral.ai/news/announcing-mistral-7b/",
       "preprompt": "",
       "chatPromptTemplate" : "{{#each messages}}{{#ifUser}}User: {{content}}{{/ifUser}}<end_of_utterance>\nAssistant: {{#ifAssistant}}{{content}}{{/ifAssistant}}{{/each}}",
+      "multimodal" : true,
       "parameters": {
         "temperature": 0.8,
         "top_p": 0.95,
@@ -137,7 +138,7 @@ MODELS=`[
         "truncate": 1000,
         "max_new_tokens": 1024,
         "do_sample" : true,
-        "stop_sequences": ["<end_of_utterance>", "\nUser:"]
+        "stop_sequences": ["<end_of_utterance>", "User:"]
       },
       "promptExamples": [
         {
